tests/pgsql: adjust pcap_cnt for request-tx
Termination messages in PGSQL are unidirectional transactions from the
client-side only. This means that, once we start to track tx progress
per-direction, these may take longer to be considered completed by the
engine, as we won't mark the tx as fully completed when this request is
seen.

Related to
Bug #7113